% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inception_resnet_v1.R
\name{load_weights}
\alias{load_weights}
\title{Load Pretrained Weights}
\usage{
load_weights(mdl, name, cache_dir = ".")
}
\arguments{
\item{mdl}{{torch nn_module} -- torch model.}

\item{name}{{str} -- Must be either 'vggface2' or 'casia-webface'. Name of dataset that was used to generate
pretrained state_dict.}

\item{cache_dir}{(NULL or str) dir to cache weights. If NULL, disable cache. Defaul: '.' (working directory)}
}
\description{
Download pretrained state_dict and load into model.
}
