% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcnn.R
\name{nn_onet}
\alias{nn_onet}
\title{MTCNN ONet.}
\usage{
nn_onet(pretrained = TRUE)
}
\arguments{
\item{pretrained}{{bool} Whether or not to load saved pretrained weights  (default: {TRUE})}
}
\description{
MTCNN ONet.
}
