% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcnn.R
\name{nn_pnet}
\alias{nn_pnet}
\title{MTCNN PNet.}
\usage{
nn_pnet(pretrained = TRUE)
}
\arguments{
\item{pretrained}{{bool} Whether or not to load saved pretrained weights  (default: {TRUE})}
}
\description{
MTCNN PNet.
}
