% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcnn.R
\name{nn_rnet}
\alias{nn_rnet}
\title{MTCNN RNet.}
\usage{
nn_rnet(pretrained = TRUE)
}
\arguments{
\item{pretrained}{{bool} Whether or not to load saved pretrained weights  (default: {TRUE})}
}
\description{
MTCNN RNet.
}
